pyproject.toml
[project]
name = "icon-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flet>=0.24.1",
    "open-clip-torch[training]>=2.28.0",
    "pillow>=11.0.0",
    "requests>=2.32.3",
    "torch>=2.5.0",
    "voyager>=2.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
ignore = [
    "ANN101", # missing type self
    "ANN102", # missing type cls
    "TD001",  # invalid todo tag
    "TD002",  # missing todo author
    "TD003",  # missing todo link
    "D1",     # missing docstring
    "D203",   # one blank line before class
    "D213",   # multi-line summary second line
]

# logger-objects = ["src.library.logger.LOGGER"]
select = ["ALL"]
unfixable = [
    "F401",   # unused import
    "F841",   # unused variable
    "ERA001", # commented out code
]

[tool.ruff.lint.pycodestyle]
max-line-length = 102

# [tool.ruff.lint.pylint]
# max-args = 8
